rename: index.tsx -> component.tsx 형식으로 변경 및 동일 폴더 내 경로 수정
