+++ src/pages/index.ts
@@ -1,1 +1,1 @@
-export { GatheringListPage as default } from '@/pages/gatheringListPage';
+export { GatheringListPage as default } from './gatheringListPage/GatheringListPage';
